Adição de classes , refatoração de rotas e documentação
