+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "node ./build/index.js",
     "build": "tsc",
-    "dev": "ts-node-dev ./src/index.ts",
+    "dev": "ts-node-dev ./src/server.ts",
     "migrations": "tsc && node ./build/database/migrations.js"
   },
   "keywords": [],
