+++ src/models/Turma.ts
@@ -7,11 +7,15 @@ export type TTurma = {
 }
 
 export class Turma {
+    private id?: string
     constructor(
-        private id: string,
         private nome: string,
         private modulo: number
-    ){}
+    ){
+        if(!this.id){
+            this.id = this.id = String(Date.now())
+        }
+    }
 
     public getId(){
         return this.id
