+++ src/server.ts
@@ -1,0 +1,11 @@
+import app from "./app";
+import { AddressInfo } from "net";
+
+const server = app.listen(3003, () => {
+  if (server) {
+    const address = server.address() as AddressInfo;
+    console.log(`Server is running in http://localhost:${address.port}`);
+  } else {
+    console.error(`Failure upon starting server.`);
+  }
+});
